centralize builder in common emu file
